refactor: refatora todo o codigo, deixando-o mais limpo e organizado de acordo com a arquitetura rest
