+++ src/main/java/br/com/actionlabs/carboncalc/adapter/controller/dto/request/calculation/CarbonCalculationResultDTO.java
@@ -1,4 +1,4 @@
-package br.com.actionlabs.carboncalc.dto;
+package br.com.actionlabs.carboncalc.adapter.controller.dto.request.calculation;
 
 import lombok.Data;
