+++ src/main/java/br/com/actionlabs/carboncalc/adapter/controller/dto/request/server/ServerStatusDTO.java
@@ -1,4 +1,4 @@
-package br.com.actionlabs.carboncalc.dto;
+package br.com.actionlabs.carboncalc.adapter.controller.dto.request.server;
 
 public record ServerStatusDTO(String version, long timestamp, String timestampStr) {
