+++ src/main/java/br/com/actionlabs/carboncalc/adapter/controller/dto/response/calculation/StartCalcResponseDTO.java
@@ -1,4 +1,4 @@
-package br.com.actionlabs.carboncalc.dto;
+package br.com.actionlabs.carboncalc.adapter.controller.dto.response.calculation;
 
 import lombok.Data;
