+++ src/main/java/br/com/actionlabs/carboncalc/core/UseCase.java
@@ -1,0 +1,5 @@
+package br.com.actionlabs.carboncalc.core;
+
+public interface UseCase<I, O> {
+    O execute(I input);
+}
