+++ src/main/java/br/com/actionlabs/carboncalc/core/domain/solidWasteEmission/SolidWasteEmissionFactor.java
@@ -1,4 +1,4 @@
-package br.com.actionlabs.carboncalc.model;
+package br.com.actionlabs.carboncalc.core.domain.solidWasteEmission;
 
 import lombok.Data;
 import org.springframework.data.annotation.Id;
