+++ src/main/java/br/com/actionlabs/carboncalc/core/domain/transportationEmission/TransportationEmissionFactor.java
@@ -1,6 +1,5 @@
-package br.com.actionlabs.carboncalc.model;
+package br.com.actionlabs.carboncalc.core.domain.transportationEmission;
 
-import br.com.actionlabs.carboncalc.enums.TransportationType;
 import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
