+++ src/main/java/br/com/actionlabs/carboncalc/core/repository/EnergyEmissionFactorRepository.java
@@ -1,7 +1,7 @@
-package br.com.actionlabs.carboncalc.repository;
+package br.com.actionlabs.carboncalc.core.repository;
 
 
-import br.com.actionlabs.carboncalc.model.EnergyEmissionFactor;
+import br.com.actionlabs.carboncalc.core.domain.energyEmission.EnergyEmissionFactor;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
