+++ src/main/java/br/com/actionlabs/carboncalc/core/repository/SolidWasteEmissionFactorRepository.java
@@ -1,7 +1,6 @@
-package br.com.actionlabs.carboncalc.repository;
+package br.com.actionlabs.carboncalc.core.repository;
 
-import br.com.actionlabs.carboncalc.model.EnergyEmissionFactor;
-import br.com.actionlabs.carboncalc.model.SolidWasteEmissionFactor;
+import br.com.actionlabs.carboncalc.core.domain.solidWasteEmission.SolidWasteEmissionFactor;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
