+++ src/main/java/br/com/actionlabs/carboncalc/core/voter/Voter.java
@@ -1,0 +1,5 @@
+package br.com.actionlabs.carboncalc.core.voter;
+
+public interface Voter<I> {
+    void vote(I useCaseInput);
+}
