--- src/main/java/br/com/actionlabs/carboncalc/dto/TransportationDTO.java
@@ -1,10 +1,0 @@
-package br.com.actionlabs.carboncalc.dto;
-
-import br.com.actionlabs.carboncalc.enums.TransportationType;
-import lombok.Data;
-
-@Data
-public class TransportationDTO {
-  private TransportationType type;
-  private int monthlyDistance;
-}
